+++ .gitmodules
@@ -1,9 +1,9 @@
 [submodule "tests_integration/definitions/conda_forge_pinning/resources/feedstock"]
-	path = tests_integration/definitions/conda_forge_pinning/resources/feedstock
+	path = tests_integration/lib/_definitions/conda_forge_pinning/resources/feedstock
 	url = https://github.com/conda-forge/conda-forge-pinning-feedstock.git
 [submodule "tests_integration/definitions/pydantic/resources/feedstock"]
-	path = tests_integration/definitions/pydantic/resources/feedstock
+	path = tests_integration/lib/_definitions/pydantic/resources/feedstock
 	url = https://github.com/conda-forge/pydantic-feedstock.git
 [submodule "tests_integration/definitions/pydantic/resources/feedstock_v1"]
-	path = tests_integration/definitions/pydantic/resources/feedstock_v1
+	path = tests_integration/lib/_definitions/pydantic/resources/feedstock_v1
 	url = https://github.com/conda-forge/pydantic-feedstock.git
