--- tests_integration/definitions/__init__.py
@@ -1,10 +1,0 @@
-from tests_integration.definitions import conda_forge_pinning, pydantic
-from tests_integration.lib import TestCase
-
-TEST_CASE_MAPPING: dict[str, list[TestCase]] = {
-    "conda-forge-pinning": conda_forge_pinning.ALL_TEST_CASES,
-    "pydantic": pydantic.ALL_TEST_CASES,
-}
-"""
-Maps from feedstock name to a list of all test cases for that feedstock.
-"""
