+++ tests_integration/lib/_definitions/conda_forge_pinning/__init__.py
@@ -2,18 +2,18 @@ from pathlib import Path
 
 from fastapi import APIRouter
 
-from tests_integration.lib import IntegrationTestHelper, TestCase
+from ..base_classes import AbstractIntegrationTestHelper, TestCase
 
 
 class SetupPinnings(TestCase):
     def get_router(self) -> APIRouter:
         return APIRouter()
 
-    def prepare(self, helper: IntegrationTestHelper):
+    def prepare(self, helper: AbstractIntegrationTestHelper):
         feedstock_dir = Path(__file__).parent / "resources" / "feedstock"
         helper.overwrite_feedstock_contents("conda-forge-pinning", feedstock_dir)
 
-    def validate(self, helper: IntegrationTestHelper):
+    def validate(self, helper: AbstractIntegrationTestHelper):
         pass
